+++ lib/repository.rb
@@ -2,7 +2,7 @@
 require_relative './repository_helper_modules/helper'
 # helper module for all repositories
 module Repository
-  include InvoiceItemHelper
+  include Helper
   def inspect
     "#<#{self.class} #{@repository.size} rows>"
   end
