+++ lib/repository_helper_modules/helper.rb
@@ -1,1 +1,7 @@
-require_relative './invoice_item_helper'
+require_relative './invoice_helper.rb'
+require_relative './invoice_item_helper.rb'
+
+module Helper
+  include InvoiceHelper
+  include InvoiceItemHelper
+end
